requirements.txt
Django==1.8.4
djangorestframework==3.2.4
djangorestframework-jwt==1.7.2
MySQL-python==1.2.5
PyJWT==1.4.0
pytz==2015.7
